query_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.5.0
qdrant-client==1.9.1
pydantic==2.10.6
streamlit==1.43.0
requests==2.32.3